# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Node Comics API",
    "description": "私有漫画图片、持久化翻译任务、普通与 PLUS 会员权益及周期页数额度。",
    "version": "0.3.0"
  },
  "paths": {
...
    "/v1/translation-submissions": {
      "post": {
        "tags": [
          "submissions"
        ],
        "summary": "Submit",
        "operationId": "submit_v1_translation_submissions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "submissions"
        ],
        "summary": "Submissions",
        "operationId": "submissions_v1_translation_submissions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/translation-submissions/{submission_id}": {
      "get": {
        "tags": [
          "submissions"
        ],
        "summary": "Submission Get",
        "operationId": "submission_get_v1_translation_submissions__submission_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Submission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/translation-submissions/{submission_id}/cancel": {
      "post": {
        "tags": [
          "submissions"
        ],
        "summary": "Submission Cancel",
        "operationId": "submission_cancel_v1_translation_submissions__submission_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Submission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/uploads/{upload_id}/content": {
      "put": {
        "tags": [
          "submissions"
        ],
        "summary": "Upload Content",
        "operationId": "upload_content_v1_uploads__upload_id__content_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "upload_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Upload Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/uploads/{upload_id}/complete": {
      "post": {
        "tags": [
          "submissions"
        ],
        "summary": "Upload Complete",
        "operationId": "upload_complete_v1_uploads__upload_id__complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "upload_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Upload Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ItemRequest": {
        "properties": {
          "client_item_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Client Item Id"
          },
          "image_sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Image Sha256"
          },
          "byte_size": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Byte Size"
          },
          "content_type": {
            "type": "string",
            "maxLength": 60,
            "title": "Content Type"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name",
            "default": ""
          },
          "file_hash": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-f0-9]{64}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Hash"
          },
          "page_index": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Index"
          },
          "asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "client_item_id",
          "image_sha256",
          "byte_size",
          "content_type"
        ],
        "title": "ItemRequest"
      },
...
      "SubmissionRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "classic",
              "redraw"
            ],
            "title": "Mode"
          },
          "target_language": {
            "type": "string",
            "title": "Target Language"
          },
          "max_quota_pages": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Max Quota Pages"
          },
          "expected_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Kind"
          },
          "regenerate": {
            "type": "boolean",
            "title": "Regenerate",
            "default": false
          },
          "rerun_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rerun Job Id"
          },
          "acknowledge_unknown_cost": {
            "type": "boolean",
            "title": "Acknowledge Unknown Cost",
            "default": false
          },
          "reading_session_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading Session Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemRequest"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "target_language",
          "max_quota_pages",
          "items"
        ],
        "title": "SubmissionRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}